requirements.txt
azure-ai-formrecognizer
azure-core